template_engine v0.0.1 TemplateEngine

Functions

parse(template)

scan(string, token)

Specs

scan(String.t, String.t) :: {String.t, String.t, String.t}

Scans for token in string.

When token was found, it returns a tuple {prefix, token, postfix}, where prefix is the string before the token, token is the token itself and postfix is the string after the token.

If token was not found, the entire string is treated as the prefix and {string, "", ""} is returned.